Trivial prototype change (Estrndup).
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@27969 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e16/e/src/memory.c
@@ -62,7 +62,7 @@ Estrdup(const char *s)
 }
 
 char               *
-Estrndup(const char *s, int n)
+Estrndup(const char *s, size_t n)
 {
 #if USE_LIBC_STRNDUP
    if (s)
